Fixed weird bug in reading data from tcpclient
In the MATLAB tcpclient.read() function, if the number of bytes to
read is passed as a uint16 value, the BytesAvailable value is then
converted into a uint16 value which has a maximum of 65535. This was
causing issues for reading 10^4 data points (120000 uint8 values).
Changed code so that headerLength is now a double value.